+++ nemo_cmd/combine.py
@@ -15,23 +15,21 @@
 """SalishSeaCmd command plug-in for combine sub-command.
 
 Combine per-processor files from an MPI Salish Sea NEMO run into single
-files with the same name-root and move them to a specified directory.
+files with the same name-root.
 """
-import glob
-import gzip
 import logging
-import math
-import multiprocessing
 import os
+import shlex
+from pathlib import Path
 import shutil
 import subprocess
-import sys
 
 import cliff.command
+import yaml
 
 from nemo_cmd import lib
 
-log = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
 
 class Combine(cliff.command.Command):
@@ -41,14 +39,17 @@ class Combine(cliff.command.Command):
     def get_parser(self, prog_name):
         parser = super(Combine, self).get_parser(prog_name)
         parser.description = '''
-            Combine the per-processor results files from an MPI
-            Salish Sea NEMO run described in DESC_FILE
-            into files in RESULTS_DIR.
+            Combine the per-processor results and/or restart files from an MPI
+            NEMO run described in DESC_FILE using the the NEMO rebuild_nemo
+            tool.
             Delete the per-processor files.
-
-            If RESULTS_DIR does not exist it will be created.
         '''
-        lib.add_combine_gather_options(parser)
+        parser.add_argument(
+            'run_desc_file',
+            type=Path,
+            metavar='RUN_DESC_FILE',
+            help='file path/name of run description YAML file'
+        )
         return parser
 
     def take_action(self, parsed_args):
@@ -59,127 +60,88 @@ class Combine(cliff.command.Command):
 
         The output of `rebuild_nemo` for each file set is logged
         at the INFO level.
-        The combined results files that `rebuild_nemo` produces are moved
-        to the directory given by `parsed_args.results_dir`.
         """
-        run_desc = lib.load_run_desc(parsed_args.desc_file)
-        n_processors = lib.get_n_processors(run_desc)
-        name_roots = _get_results_files(parsed_args)
-        if name_roots:
-            rebuild_nemo_script = _find_rebuild_nemo_script(run_desc)
-            _combine_results_files(
-                rebuild_nemo_script, name_roots, n_processors
-            )
-            os.remove('nam_rebuild')
-        _netcdf4_deflate_results()
-        if name_roots:
-            _move_results(name_roots, parsed_args.results_dir)
-            _compress_results(name_roots, parsed_args)
-            _delete_results_files(name_roots, parsed_args)
+        combine(parsed_args.run_desc_file)
 
 
-def _find_rebuild_nemo_script(run_desc):
-    nemo_code_repo = os.path.abspath(
-        os.path.expandvars(os.path.expanduser(run_desc['paths']['NEMO-code']))
-    )
-    rebuild_nemo_exec = os.path.join(
-        nemo_code_repo, 'NEMOGCM', 'TOOLS', 'REBUILD_NEMO', 'rebuild_nemo.exe'
-    )
-    if not os.path.lexists(rebuild_nemo_exec):
-        log.error(
-            '{} not found - did you forget to build it?'
-            .format(rebuild_nemo_exec)
-        )
-        sys.exit(2)
-    rebuild_nemo_script = os.path.splitext(rebuild_nemo_exec)[0]
-    return rebuild_nemo_script
+def combine(run_desc_file):
+    """Run the NEMO :program:`rebuild_nemo` tool for each set of
+    per-processor results files.
+
+    The output of :program:`rebuild_nemo` for each file set is logged
+    at the INFO level.
+
+    :param run_desc_file: File path/name of the run description YAML file.
+    :type run_desc_file: :py:class:`pathlib.Path`
+    """
+    with run_desc_file.open('rt') as f:
+        run_desc = yaml.safe_load(f)
+    name_roots = _get_results_files()
+    if name_roots:
+        rebuild_nemo_script = _find_rebuild_nemo_script(run_desc)
+        _combine_results_files(rebuild_nemo_script, name_roots)
+        _delete_results_files(name_roots)
 
 
-def _get_results_files(args):
-    if args.delete_restart:
-        restart_pattern = '*_restart_[0-9][0-9][0-9][0-9].nc'
-        for fn in glob.glob(restart_pattern):
-            os.remove(fn)
+def _get_results_files():
     result_pattern = '*_0000.nc'
-    name_roots = [fn[:-8] for fn in glob.glob(result_pattern)]
+    name_roots = [str(fn)[:-8] for fn in Path.cwd().glob(result_pattern)]
     if not name_roots:
-        log.info(
+        logger.info(
             'no files found that match the {} pattern'.format(result_pattern)
         )
     return name_roots
 
 
-def _combine_results_files(rebuild_nemo_script, name_roots, n_processors):
+def _find_rebuild_nemo_script(run_desc):
+    nemo_code_config = Path(
+        os.path.expandvars(run_desc['paths']['NEMO code config'])
+    ).expanduser().resolve()
+    rebuild_nemo_exec = (
+        nemo_code_config / '..' / 'TOOLS' / 'REBUILD_NEMO' / 'rebuild_nemo.exe'
+    )
+    if not rebuild_nemo_exec.exists():
+        logger.error(
+            '{} not found - did you forget to build it?'
+            .format(rebuild_nemo_exec)
+        )
+        raise SystemExit
+    rebuild_nemo_script = rebuild_nemo_exec.with_suffix('').resolve()
+    return rebuild_nemo_script
+
+
+def _combine_results_files(rebuild_nemo_script, name_roots):
     for fn in name_roots:
-        nfiles = len(glob.glob('{fn}_[0-9][0-9][0-9][0-9].nc'.format(fn=fn)))
+        files = Path.cwd().glob('{fn}_[0-9][0-9][0-9][0-9].nc'.format(fn=fn))
+        # Count the number of items yielded by the glob generator
+        nfiles = sum(1 for _ in files)
         if nfiles == 1:
+            # Results from a single processor are simply renamed
             shutil.move('{fn}_0000.nc'.format(fn=fn), '{fn}.nc'.format(fn=fn))
-            log.info('{fn}_0000.nc renamed to {fn}.nc'.format(fn=fn))
-        elif nfiles < n_processors:
-            log.info('{fn}_*.nc not combined'.format(fn=fn))
+            logger.info('{fn}_0000.nc renamed to {fn}.nc'.format(fn=fn))
         else:
+            cmd = (
+                '{rebuild_nemo_script} {fn} {nfiles}'.format(
+                    rebuild_nemo_script=rebuild_nemo_script,
+                    fn=fn,
+                    nfiles=nfiles
+                )
+            )
+            logger.info(cmd)
             result = subprocess.check_output(
-                [rebuild_nemo_script, fn, str(n_processors)],
+                shlex.split(cmd),
                 stderr=subprocess.STDOUT,
                 universal_newlines=True
             )
-            log.info(result)
+            logger.info(result)
+            os.unlink('nam_rebuild')
 
 
-def _netcdf4_deflate_results():
-    log.info('Starting netCDF4 deflation...')
-    patterns = (
-        '*_restart.nc', '*_restart_trc.nc', '*_grid_[TUVW].nc', '*_ptrc_T.nc'
-    )
-    for pattern in patterns:
-        for fn in glob.glob(pattern):
-            result = lib.netcdf4_deflate(
-                fn, math.floor(multiprocessing.cpu_count() / 2)
-            )
-            if result:
-                log.info(result)
-            else:
-                log.info('netCDF4 deflated {}'.format(fn))
-
-
-def _move_results(name_roots, results_dir):
-    abs_results_dir = os.path.abspath(results_dir)
-    if os.path.exists(abs_results_dir):
-        if os.path.samefile(os.getcwd(), abs_results_dir):
-            return
-    else:
-        os.makedirs(abs_results_dir)
-    postfix = '' if results_dir.endswith('/') else '/'
-    for fn in _results_files(name_roots):
-        log.info('moving {} to {}{}'.format(fn, results_dir, postfix))
-        shutil.move(os.path.join('.', fn), os.path.join(abs_results_dir, fn))
-
-
-def _results_files(name_roots):
-    for fn in ('.'.join((n, 'nc')) for n in name_roots):
-        yield fn
-
-
-def _compress_results(name_roots, args):
-    if not args.compress:
-        return
-    log.info('Starting compression...')
-    for fn in _results_files(name_roots):
-        if 'restart' in fn and not args.compress_restart:
-            continue
-        fp = os.path.join(args.results_dir, fn)
-        with open(fp, 'rb') as f_in:
-            fpgz = '.'.join((fp, 'gz'))
-            with gzip.open(fpgz, 'wb') as f_out:
-                f_out.writelines(f_in)
-        os.remove(fp)
-        log.info('compressed {}'.format(fp))
-
-
-def _delete_results_files(name_roots, args):
-    if args.keep_proc_results:
-        return
-    log.info('Deleting per-processor files...')
+def _delete_results_files(name_roots):
+    logger.info('Deleting per-processor files...')
     for name_root in name_roots:
-        for fn in glob.glob(name_root + '_[0-9][0-9][0-9][0-9].nc'):
-            os.remove(fn)
+        filepaths = Path.cwd().glob(
+            '{name_root}_[0-9][0-9][0-9][0-9].nc'.format(name_root=name_root)
+        )
+        for fp in filepaths:
+            fp.unlink()