+++ tests/test_api.py
@@ -16,15 +16,10 @@
 """
 from io import StringIO
 try:
-    from unittest.mock import (
-        Mock,
-        patch,
-    )
+    from unittest.mock import Mock, patch
 except ImportError:
-    from mock import (
-        Mock,
-        patch,
-    )
+    # Python 2.7
+    from mock import Mock, patch
 
 import cliff.app
 import cliff.command
@@ -32,89 +27,6 @@ import pytest
 import yaml
 
 import nemo_cmd.api
-
-
-class TestCombine(object):
-    @patch('nemo_cmd.api._run_subcommand')
-    def test_combine_default_args(self, m_run_subcommand):
-        app, app_args = Mock(spec=cliff.app.App), []
-        nemo_cmd.api.combine(app, app_args, 'run_desc_file', 'results_dir')
-        m_run_subcommand.assert_called_once_with(
-            app, app_args, ['combine', 'run_desc_file', 'results_dir']
-        )
-
-    @patch('nemo_cmd.api._run_subcommand')
-    def test_combine_keep_proc_results_arg(self, m_run_subcommand):
-        app, app_args = Mock(spec=cliff.app.App), []
-        nemo_cmd.api.combine(
-            app,
-            app_args,
-            'run_desc_file',
-            'results_dir',
-            keep_proc_results=True
-        )
-        m_run_subcommand.assert_called_once_with(
-            app, app_args,
-            ['combine', 'run_desc_file', 'results_dir', '--keep-proc-results']
-        )
-
-    @patch('nemo_cmd.api._run_subcommand')
-    def test_combine_no_compress_arg(self, m_run_subcommand):
-        app, app_args = Mock(spec=cliff.app.App), []
-        nemo_cmd.api.combine(
-            app, app_args, 'run_desc_file', 'results_dir', no_compress=True
-        )
-        m_run_subcommand.assert_called_once_with(
-            app, app_args,
-            ['combine', 'run_desc_file', 'results_dir', '--no-compress']
-        )
-
-    @patch('nemo_cmd.api._run_subcommand')
-    def test_combine_compress_restart_arg(self, m_run_subcommand):
-        app, app_args = Mock(spec=cliff.app.App), []
-        nemo_cmd.api.combine(
-            app,
-            app_args,
-            'run_desc_file',
-            'results_dir',
-            compress_restart=True
-        )
-        m_run_subcommand.assert_called_once_with(
-            app, app_args,
-            ['combine', 'run_desc_file', 'results_dir', '--compress-restart']
-        )
-
-    @patch('nemo_cmd.api._run_subcommand')
-    def test_combine_delete_restart_arg(self, m_run_subcommand):
-        app, app_args = Mock(spec=cliff.app.App), []
-        nemo_cmd.api.combine(
-            app, app_args, 'run_desc_file', 'results_dir', delete_restart=True
-        )
-        m_run_subcommand.assert_called_once_with(
-            app, app_args,
-            ['combine', 'run_desc_file', 'results_dir', '--delete-restart']
-        )
-
-    @patch('nemo_cmd.api._run_subcommand')
-    def test_combine_all_args(self, m_run_subcommand):
-        app, app_args = Mock(spec=cliff.app.App), []
-        nemo_cmd.api.combine(
-            app,
-            app_args,
-            'run_desc_file',
-            'results_dir',
-            keep_proc_results=True,
-            no_compress=True,
-            compress_restart=True,
-            delete_restart=True
-        )
-        m_run_subcommand.assert_called_once_with(
-            app, app_args, [
-                'combine', 'run_desc_file', 'results_dir',
-                '--keep-proc-results', '--no-compress', '--compress-restart',
-                '--delete-restart'
-            ]
-        )
 
 
 class TestRunDescription(object):