+++ tests/test_combine.py
@@ -14,10 +14,25 @@
 # limitations under the License.
 """SalishSeaCmd combine sub-command plug-in unit tests
 """
+import shlex
+from pathlib import Path
+
+import subprocess
+
 try:
-    from unittest.mock import Mock, patch
+    from types import SimpleNamespace
 except ImportError:
-    from mock import Mock, patch
+    # Python 2.7
+    class SimpleNamespace:
+        def __init__(self, **kwargs):
+            self.__dict__.update(kwargs)
+
+
+try:
+    from unittest.mock import call, Mock, patch
+except ImportError:
+    # Python 2.7
+    from mock import call, Mock, patch
 
 import cliff.app
 import pytest
@@ -31,137 +46,115 @@ def combine_cmd():
     return nemo_cmd.combine.Combine(Mock(spec=cliff.app.App), [])
 
 
-class TestGetParser:
-    """Unit tests for `salishsea combine` sub-command command-line parser.
+class TestParser:
+    """Unit tests for `nemo combine` sub-command command-line parser.
     """
 
     def test_get_parser(self, combine_cmd):
-        parser = combine_cmd.get_parser('salishsea combine')
-        assert parser.prog == 'salishsea combine'
+        parser = combine_cmd.get_parser('nemo combine')
+        assert parser.prog == 'nemo combine'
+
+    def test_parsed_args(self, combine_cmd):
+        parser = combine_cmd.get_parser('nemo combine')
+        parsed_args = parser.parse_args(['nemo.yaml'])
+        assert parsed_args.run_desc_file == Path('nemo.yaml')
 
 
-@patch('nemo_cmd.combine.os.path.lexists')
-class TestFindRebuildNemoScript:
-    @pytest.mark.parametrize(
-        'nemo_path, expected', [
-            ('NEMO-code', 'NEMO-code/NEMOGCM/TOOLS/REBUILD_NEMO/rebuild_nemo'),
-        ]
-    )
-    @patch('nemo_cmd.combine.os.path.abspath')
-    def test_find_rebuild_nemo_script_found(
-        self,
-        mock_abspath,
-        mock_lexists,
-        nemo_path,
-        expected,
-    ):
-        """_find_rebuild_nemo_exec returns script name if executable exists
-        """
-        run_desc = {'paths': {'NEMO-code': 'NEMO-code'}}
-        mock_lexists.return_value = True
-        mock_abspath.return_value = nemo_path
-        script = nemo_cmd.combine._find_rebuild_nemo_script(run_desc)
-        assert script == 'NEMO-code/NEMOGCM/TOOLS/REBUILD_NEMO/rebuild_nemo'
+class TestTakeAction:
+    """Unit test for `nemo combine` sub-command take_action() method.
+    """
 
-    @patch('nemo_cmd.combine.log.error')
-    def test_find_rebuild_nemo_script_not_found(self, mock_log, mock_lexists):
-        """_find_rebuild_nemo_exec logs error if executable not found
-        """
-        run_desc = {'paths': {'NEMO-code': 'NEMO-code'}}
-        mock_lexists.return_value = False
-        with pytest.raises(SystemExit):
-            nemo_cmd.combine._find_rebuild_nemo_script(run_desc)
-        assert mock_log.called
+    @patch('nemo_cmd.combine.combine')
+    def test_take_action(self, m_combine, combine_cmd):
+        parsed_args = SimpleNamespace(run_desc_file=Path('nemo.yaml'))
+        combine_cmd.take_action(parsed_args)
+        m_combine.assert_called_once_with(Path('nemo.yaml'))
 
 
+@patch('nemo_cmd.combine.Path')
+@patch('nemo_cmd.combine.logger')
 class TestGetResultsFiles:
-    @patch('nemo_cmd.combine.glob.glob')
-    def test_get_results_files(self, mock_glob):
-        """_get_results_files returns list of name-roots and count of files
-        """
-        mock_glob.side_effect = (['foo_0000.nc', 'bar_0000.nc'],
-                                 ['foo_0000.nc', 'foo_0001.nc', 'foo_0002.nc'])
-        args = Mock(delete_restart=False)
-        name_roots = nemo_cmd.combine._get_results_files(args)
+    def test_get_results_files(self, m_logger, m_path):
+        m_path.cwd().glob.return_value = ['foo_0000.nc', 'bar_0000.nc']
+        name_roots = nemo_cmd.combine._get_results_files()
         assert name_roots == ['foo', 'bar']
 
-    @patch('nemo_cmd.combine.log.info')
-    def test_get_results_files_none_found(self, mock_log):
-        """_get_results_files logs info message if no results files exists
-        """
-        args = Mock(delete_restart=False)
-        nemo_cmd.combine._get_results_files(args)
-        assert mock_log.called
+    def test_get_results_files_none_found(self, m_logger, m_path):
+        nemo_cmd.combine._get_results_files()
+        assert m_logger.info.called
 
-    @patch('nemo_cmd.combine.glob.glob')
-    @patch('nemo_cmd.combine.os.remove')
-    def test_get_results_files_delete_restart(self, mock_rm, mock_glob):
-        """_get_results_files deletes restart files
-        """
-        mock_glob.side_effect = (
-            ['baz_restart_0000.nc', 'baz_restart_0001.nc'],
-            ['foo_0000.nc', 'bar_0000.nc'],
-            ['foo_0000.nc', 'foo_0001.nc', 'foo_0002.nc']
+
+@patch('nemo_cmd.combine.logger')
+class TestFindRebuildNemoScript:
+    """Unit tests for _find_rebuild_nemo_script function.
+    """
+
+    @pytest.mark.parametrize(
+        'nemo_code_config', [
+            'NEMO-3.6-code/NEMOGCM/CONFIG/',
+            'NEMO-3.6/CONFIG/',
+        ]
+    )
+    def test_find_rebuild_nemo_script(
+        self, m_logger, nemo_code_config, tmpdir
+    ):
+        nemo_code_config = tmpdir.ensure_dir(nemo_code_config)
+        nemo_code_config.ensure('../TOOLS/REBUILD_NEMO/rebuild_nemo.exe')
+        script_path = nemo_code_config.ensure(
+            '../TOOLS/REBUILD_NEMO/rebuild_nemo'
         )
-        args = Mock(delete_restart=True)
-        nemo_cmd.combine._get_results_files(args)
-        assert mock_rm.call_count == 2
+        run_desc = {'paths': {'NEMO code config': str(nemo_code_config)}}
+        rebuild_nemo_script = nemo_cmd.combine._find_rebuild_nemo_script(
+            run_desc
+        )
+        assert rebuild_nemo_script == Path(str(script_path))
+
+    @pytest.mark.parametrize(
+        'nemo_code_config', [
+            'NEMO-3.6-code/NEMOGCM/CONFIG/',
+            'NEMO-3.6/CONFIG/',
+        ]
+    )
+    def test_no_rebuild_nemo_script(self, m_logger, nemo_code_config, tmpdir):
+        nemo_code_config = tmpdir.ensure_dir(nemo_code_config)
+        run_desc = {'paths': {'NEMO code config': str(nemo_code_config)}}
+        with pytest.raises(SystemExit):
+            nemo_cmd.combine._find_rebuild_nemo_script(run_desc)
+        assert m_logger.error.called
 
 
+@patch('nemo_cmd.combine.logger')
+@patch('nemo_cmd.combine.Path')
 class TestCombineResultsFiles:
-    @patch('nemo_cmd.combine.glob.glob')
+    """Unit tests for _combine_results_files function.
+    """
+
+    @patch('nemo_cmd.combine.shutil.move')
+    def test_single_processor_result(self, m_move, m_path, m_logger):
+        m_path.cwd().glob.return_value = ['foo_0000.nc']
+        nemo_cmd.combine._combine_results_files('rebuild_nemo', ['foo'])
+        assert m_move.call_args == call('foo_0000.nc', 'foo.nc')
+
     @patch('nemo_cmd.combine.subprocess.check_output')
-    def test_combine_results_files(self, mock_chk_out, mock_glob):
-        """_combine_results_files calls subprocess.check_output for each name-root
-        """
-        mock_glob.side_effect = (
-            ['foo_0000.nc', 'foo_0001.nc', 'foo_0002.nc'],
-            ['bar_0000.nc', 'bar_0001.nc', 'bar_0002.nc'],
+    @patch('nemo_cmd.combine.os.unlink')
+    def test_rebuild_nemo_subprocess(
+        self, m_unlink, m_check_output, m_path, m_logger
+    ):
+        m_path.cwd().glob.return_value = ['foo_0000.nc', 'foo_0001.nc']
+        nemo_cmd.combine._combine_results_files('rebuild_nemo', ['foo'])
+        assert m_check_output.call_args == call(
+            shlex.split('rebuild_nemo foo 2'),
+            stderr=subprocess.STDOUT,
+            universal_newlines=True
         )
-        nemo_cmd.combine._combine_results_files(
-            'rebuild_nemo', ['foo', 'bar'], 3
-        )
-        assert mock_chk_out.call_count == 2
 
 
-class TestMoveResults:
-    @patch('nemo_cmd.combine.shutil.move')
-    def test_move_results_pwd(self, mock_move):
-        """_move_results does nothing if results_dir is pwd
-        """
-        nemo_cmd.combine._move_results(['foo'], './')
-        assert not mock_move.called
+@patch('nemo_cmd.combine.Path')
+class TestDeleteResultsFiles:
+    """Unit tests for _delete_results_files function.
+    """
 
-    @patch('nemo_cmd.combine.shutil.move')
-    @patch('nemo_cmd.combine.os.makedirs')
-    def test_move_results_makedirs(self, mock_makedirs, mock_move):
-        """_move_results creates results_dir if it doesn't exist
-        """
-        nemo_cmd.combine._move_results(['foo', 'bar'], 'baz')
-        assert mock_makedirs.called
-
-    @patch('nemo_cmd.combine.os.makedirs')
-    @patch('nemo_cmd.combine.shutil.move')
-    def test_move_results_renames(self, mock_move, mock_makedirs):
-        """_move_results calls shutil.move for each results file
-        """
-        nemo_cmd.combine._move_results(['foo', 'bar'], 'baz')
-        assert mock_move.call_count == 2
-
-
-class TestResultsFiles:
-    def test_result_files(self):
-        """_results_files generator yields name-root with .nc appended
-        """
-        fn = next(nemo_cmd.combine._results_files(['foo', 'bar']))
-        assert fn == 'foo.nc'
-
-
-class TestCompressResults:
-    @patch('nemo_cmd.combine._results_files')
-    def test_compress_results_no_compress(self, mock_res_files):
-        """_compress_results does nothing when args.compress is False
-        """
-        args = Mock(compress=False)
-        nemo_cmd.combine._compress_results(['foo', 'bar'], args)
-        assert not mock_res_files.called
+    def test_delete_results_files(self, m_path):
+        m_path.cwd().glob.return_value = [Mock(spec=Path)]
+        nemo_cmd.combine._delete_results_files(['foo'])
+        assert m_path.cwd().glob()[0].unlink.called